Add files via upload
This is the task -03 of CODSOFT Internship . It includes to design the user interface for the ATM, including options such as withdrawing, depositing, and checking the balance. Implement methods for each option, such as withdraw(amount), deposit(amount), and checkBalance(). Create a class to represent the user's bank account, which stores the account balance. Connect the ATM class with the user's bank account class to access and modify the account balance. Validate user input to ensure it is within acceptable limits . Display appropriate messages to the user based on their chosen options and the success or failure
of their transactions.